Fix all codestyle in codebase
Adapt it to google codestyle

+++ src/ggrc_workflows/migrations/versions/20170626123211_55501ef0f634_add_workfloweditor_role.py
@@ -1,6 +1,5 @@
 # Copyright (C) 2018 Google Inc.
 # Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-
 """
 Add WorkflowEditor permission role
 
@@ -29,25 +28,24 @@ roles = sa.sql.table(
 
 
 def upgrade():
-  """Upgrade database schema and/or data, creating a new revision."""
-  op.execute(
-      roles.insert().
-      values({
-          "name": "WorkflowEditor",
-          "permissions_json": "CODE DECLARED ROLE",
-          "description": "This role grants a user permission to edit workflow "
-                         "mappings and details",
-          "created_at": _NOW,
-          "updated_at": _NOW,
-          "scope": "Workflow Implied"
-      })
-  )
+    """Upgrade database schema and/or data, creating a new revision."""
+    op.execute(roles.insert().values({
+        "name":
+        "WorkflowEditor",
+        "permissions_json":
+        "CODE DECLARED ROLE",
+        "description":
+        "This role grants a user permission to edit workflow "
+        "mappings and details",
+        "created_at":
+        _NOW,
+        "updated_at":
+        _NOW,
+        "scope":
+        "Workflow Implied"
+    }))
 
 
 def downgrade():
-  """Downgrade database schema and/or data back to the previous revision."""
-  op.execute(
-      roles.delete().where(
-          roles.c.name == "WorkflowEditor"
-      )
-  )
+    """Downgrade database schema and/or data back to the previous revision."""
+    op.execute(roles.delete().where(roles.c.name == "WorkflowEditor"))
